Delete process_inputs.py
Moved all of the logic to Runner class so all the pre-processing and routines
are in one place
