+++ src/mvesuvio/analysis_fitting.py
@@ -66,7 +66,7 @@ def extractNCPFromWorkspaces(wsFinal, ic):
 
     # Ensure shape of ncp matches data
     shape = ncpForEachMass.shape
-    assert shape[0] == ic.noOfMasses
+    assert shape[0] == ic.masses.size
     assert shape[1] == wsFinal.getNumberHistograms()
     # Final dimension can be missing last col or not
     assert (shape[2] == wsFinal.blocksize()) | (shape[2] == wsFinal.blocksize() - 1)
@@ -95,7 +95,8 @@ def calculateMantidResolutionFirstMass(IC, yFitIC, ws):
         else:
             AppendSpectra(resName, "tmp", OutputWorkspace=resName)
 
-    MaskDetectors(resName, WorkspaceIndexList=IC.maskedDetectorIdx)
+    masked_idx = [ws.spectrumInfo().isMasked(i) for i in range(ws.getNumberHistograms())]
+    MaskDetectors(resName, WorkspaceIndexList=np.flatnonzero(masked_idx))
     wsResSum = SumSpectra(InputWorkspace=resName, OutputWorkspace=resName + "_Sum")
 
     normalise_workspace(wsResSum)
@@ -127,7 +128,8 @@ def subtractAllMassesExceptFirst(ic, ws, ncpForEachMass):
 
     wsSubMass = CloneWorkspace(InputWorkspace=ws, OutputWorkspace=ws.name() + "_Mass0")
     passDataIntoWS(dataX, dataY, dataE, wsSubMass)
-    MaskDetectors(Workspace=wsSubMass, WorkspaceIndexList=ic.maskedDetectorIdx)
+    # MaskDetectors(Workspace=wsSubMass, WorkspaceIndexList=ic.maskedDetectorIdx)
+    MaskDetectors(Workspace=wsSubMass, SpectraList=ic.maskedSpecAllNo)
     SumSpectra(
         InputWorkspace=wsSubMass.name(), OutputWorkspace=wsSubMass.name() + "_Sum"
     )
